+++ tests/unit/plugin.js
@@ -3,11 +3,11 @@ define([
     '$',
     'shade'
 ], function(fixture, $) {
-    var element;
+    var $element;
 
     describe('Shade plugin', function() {
         beforeEach(function() {
-            element = $(fixture);
+            $element = $(fixture);
         });
 
         describe('binding to Zepto\'s fn', function() {
@@ -25,75 +25,91 @@ define([
         });
 
         describe('invoking shade', function() {
-            it('creates shade instance on element', function() {
-                element.shade();
+            it('creates shade instance on $element', function() {
+                $element.shade();
 
-                assert.isDefined(element.data('shade'));
+                assert.isDefined($element.data('shade'));
+
+                $element.shade('destroy');
             });
 
-            it('stores element inside instance', function() {
-                element.shade();
+            it('stores $element inside instance', function() {
+                $element.shade();
 
-                assert.isDefined(element.data('shade').$shade);
+                assert.isDefined($element.data('shade').$shade);
+
+                $element.shade('destroy');
             });
         });
 
         describe('invoking shade methods before plugin is initialized', function() {
             it('throws when not initialized', function() {
-                assert.throws(function() { element.shade('open'); });
+                assert.throws(function() { $element.shade('open'); });
             });
         });
 
         describe('invoking shade methods using the plugin interface', function() {
             it('opens a shade using the open method', function(done) {
-                element.shade({
+                $element.shade({
                     opened: function() {
-                        assert.isTrue(element.data('shade').$shade.hasClass('shade--is-open'));
+                        assert.isTrue($element.data('shade').$shade.hasClass('shade--is-open'));
+                        $element.shade('destroy');
+
                         done();
                     }
                 });
 
-                element.shade('open');
+                $element.shade('open');
             });
 
             it('closes a shade item using the close method', function(done) {
-                element.shade({
+                $element.shade({
                     opened: function() {
-                        element.shade('close');
+                        $element.shade('close');
                     },
                     closed: function() {
-                        assert.isFalse(element.data('shade').$shade.hasClass('shade--is-open'));
+                        assert.isFalse($element.data('shade').$shade.hasClass('shade--is-open'));
+                        $element.shade('destroy');
+
                         done();
                     }
                 });
 
-                element.shade('open');
+                $element.shade('open');
             });
 
             it('closes a shade by clicking on it', function(done) {
-                element.shade({
+                $element.shade({
                     opened: function() {
-                        element.data('shade').$shade.trigger('click');
+                        $element.data('shade').$shade.trigger('click');
                     },
                     closed: function() {
-                        assert.isFalse(element.data('shade').$shade.hasClass('shade--is-open'));
+                        assert.isFalse($element.data('shade').$shade.hasClass('shade--is-open'));
+                        $element.shade('destroy');
+
                         done();
                     }
                 });
 
-                element.shade('open');
+                $element.shade('open');
             });
 
             it('throws for method calls that don\'t exist', function() {
-                assert.throws(function() { element.shade().shade('noMethod'); });
+                assert.throws(function() { $element.shade().shade('noMethod'); });
+
+                $element.shade('destroy');
             });
 
             it('throws when attempting to invoke private methods', function() {
-                assert.throws(function() { element.shade().shade('_init'); });
+                assert.throws(function() { $element.shade().shade('_init'); });
+
+                $element.shade('destroy');
             });
 
             it('throws when attempting to invoke methods that aren\'t functions', function() {
-                assert.throws(function() { element.shade().shade('singleItemOpen'); });
+                assert.throws(function() { $element.shade().shade('singleItemOpen'); });
+
+                $element.shade('destroy');
             });
         });
     });