+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+1.2.0
+    - Added `destroy` function
+1.1.0
+    - Added option, `cssClass`, to specify custom CSS class
 1.0.5
     - Allows `click` to take `false` as an argument. This allows users to have Shade be fully blocking.